Version 0.1.99
Using Bram's matchparen.vim plugin from standard distribution as starting point for this version.

Print line containing matching brace with echo instead of echomsg so as to not fill message history with non-message items.

Truncate matching brace line when it's gonna be too long - this avoids the dreaded "Press ENTER or type command to continue" message that would be so disruptive to workflow.

To do this, the plugin must take the &showcmd, &ruler and &laststatus options into consideration.  If you use a custom &rulerformat, you will want to set the global variable g:MP_rulerwidth to the width of your ruler.

The variable g:MP_stmt_thresh lets the user specify how far back the plugin may scan for a statement beginning a code block.